[refactor] 관련 파일들끼리 패키지화 및 불필요한 파일 삭제
