+++ src/components/layout/Layout.jsx
@@ -1,6 +1,6 @@
 import React from "react";
 import { Outlet} from "react-router-dom";
-import Header from "./Header";
+import Header from "./header/Header";
 import Footer from "./Footer";
 import './Layout.css';
