+++ src/components/pet/list/AdoptionRecommendList.jsx
@@ -1,13 +1,13 @@
 import React, { useEffect, useState } from 'react';
 import Slider from "react-slick";
-import PetCard from "./PetCard.jsx";
+import PetCard from "../PetCard.jsx";
 import { FaChevronLeft, FaChevronRight } from 'react-icons/fa';
 import "slick-carousel/slick/slick.css";
 import "slick-carousel/slick/slick-theme.css";
 import './List.css'
-import {getAdoptionList} from "../../services/adoptionService.js";
+import {getAdoptionList} from "../../../services/adoptionService.js";
 import { useNavigate } from "react-router-dom";
-import PrimaryButton from "../common/PrimaryButton.jsx";
+import PrimaryButton from "../../common/PrimaryButton.jsx";
 import Adoption from '/src/assets/images/main/Adoption.png';
 import LostAnimal from '/src/assets/images/main/LostAnimal.png';
