+++ src/pages/adoptionAnimal/AdoptionDetail.css
