+++ src/pages/lostanimal/LostAnimalDetail.jsx
