+++ src/pages/lostanimal/LostAnimalFound.jsx
@@ -1,12 +1,17 @@
-import React from 'react';
+import React from "react";
+import "./LostAnimal.css";
 
 function LostAnimalFound() {
   return (
-    <div className="lost-animal-found">
-      <h2>발견/보호 동물 목록</h2>
+    <div className="lost-animal-page">
+      <div className="lost-animal-header">
+        <h1 className="lost-animal-title">발견/보호 동물</h1>
+        <p className="lost-animal-description">
+          발견되거나 보호 중인 반려동물입니다. 주인을 찾고 있습니다.
+        </p>
+      </div>
       <div className="lost-animal-content">
-        {/* 발견/보호 동물 목록이 들어갈 자리 */}
-        <p>발견되거나 보호 중인 동물 목록이 표시됩니다.</p>
+        {/* TODO: 발견/보호 동물 목록 컴포넌트 추가 */}
       </div>
     </div>
   );
