+++ src/pages/lostanimal/LostAnimalLost.jsx
@@ -1,12 +1,17 @@
-import React from 'react';
+import React from "react";
+import "./LostAnimal.css";
 
 function LostAnimalLost() {
   return (
-    <div className="lost-animal-lost">
-      <h2>실종 동물 목록</h2>
+    <div className="lost-animal-page">
+      <div className="lost-animal-header">
+        <h1 className="lost-animal-title">실종 동물</h1>
+        <p className="lost-animal-description">
+          실종된 반려동물을 찾고 있습니다. 아래 목록에서 실종된 반려동물을 확인하실 수 있습니다.
+        </p>
+      </div>
       <div className="lost-animal-content">
-        {/* 실종 동물 목록이 들어갈 자리 */}
-        <p>실종된 동물 목록이 표시됩니다.</p>
+        {/* TODO: 실종 동물 목록 컴포넌트 추가 */}
       </div>
     </div>
   );
