+++ src/pages/lostanimal/LostAnimalRescue.jsx
@@ -1,12 +1,17 @@
-import React from 'react';
+import React from "react";
+import "./LostAnimal.css";
 
 function LostAnimalRescue() {
   return (
-    <div className="lost-animal-rescue">
-      <h2>구조 동물 목록</h2>
+    <div className="lost-animal-page">
+      <div className="lost-animal-header">
+        <h1 className="lost-animal-title">구조 동물</h1>
+        <p className="lost-animal-description">
+          구조된 반려동물들입니다. 새로운 가족을 기다리고 있습니다.
+        </p>
+      </div>
       <div className="lost-animal-content">
-        {/* 구조 동물 목록이 들어갈 자리 */}
-        <p>구조된 동물 목록이 표시됩니다.</p>
+        {/* TODO: 구조 동물 목록 컴포넌트 추가 */}
       </div>
     </div>
   );
