+++ src/pages/mainPage/MainPage.css
