+++ src/pages/mainPage/MainPage.jsx
@@ -1,6 +1,6 @@
 import React from "react";
-import AdoptionRecommendList from "../components/pet/AdoptionRecommendList.jsx";
-import LostAnimalList from "../components/pet/LostAnimalList.jsx";
+import AdoptionRecommendList from "../../components/pet/list/AdoptionRecommendList.jsx";
+import LostAnimalList from "../../components/pet/LostAnimalList.jsx";
 import './MainPage.css';
 import Adoption from "../assets/images/main/Adoption.png";
 import LostAnimal from "../assets/images/main/LostAnimal.png";
