+++ src/pages/signup/AdditionalInfo.jsx
@@ -1,8 +1,8 @@
 // src/pages/AdditionalInfo.jsx
 import React, { useState } from 'react'
 import { useNavigate, useSearchParams } from 'react-router-dom'
-import client from '../api/client'
-import { AuthContext } from '../contexts/AuthContext'
+import client from '../../api/client'
+import { AuthContext } from '../../contexts/AuthContext'
 import { useContext, useEffect } from 'react'
 import nicknameIcon from '../assets/images/info/user.png'
 import placeholderIcon from '../assets/images/info/placeholder.png'
